Improve Part 2, don't copy entire image
Make proces_file read from the image file only one layer at a time, and
then stack it with the previous one. This way, only 2 layers should be
in memory at any point in time.

Also improve the display of the image to use a preallocated string,
instead of generating garbage with each call to .add